requirements.txt
langchain-openai 
tiktoken 
langchain-chroma
langchain.community
langchain 
GitPython 
python-dotenv

flask
jsonschema
markdown
unstructured